Rename recipe-grammar folder
