+++ packages/breadboard/src/index.ts
@@ -29,9 +29,9 @@ export {
 
 // New Syntax:
 export { Runner } from "./new/runner/runner.js";
-export { recipe, recipe as board, code } from "./new/recipe-grammar/recipe.js";
-export { addKit } from "./new/recipe-grammar/kits.js";
-export { base } from "./new/recipe-grammar/base.js";
+export { recipe, recipe as board, code } from "./new/grammar/board.js";
+export { addKit } from "./new/grammar/kits.js";
+export { base } from "./new/grammar/base.js";
 export type {
   NodeValue as NewNodeValue,
   InputValues as NewInputValues,
@@ -52,5 +52,5 @@ export type {
   ProjectBackToOutputValues as __ProjectBackToOutputValues,
   InputValues as NewInputValuesWithNodeFactory,
   AbstractValue as V,
-} from "./new/recipe-grammar/types.js";
+} from "./new/grammar/types.js";
 export { asyncGen } from "./utils/async-gen.js";
