+++ packages/breadboard/src/new/grammar/default-scope.ts
