+++ packages/breadboard/tests/new/grammar/invoke-by-await.ts
@@ -7,7 +7,7 @@
 import { z } from "zod";
 import test from "ava";
 
-import { board, code } from "../../../src/new/recipe-grammar/recipe.js";
+import { board, code } from "../../../src/new/grammar/board.js";
 import { V } from "../../../src/index.js";
 import { testKit } from "../../helpers/_test-kit.js";
