+++ packages/breadboard/tests/new/grammar/metadata.ts
@@ -8,7 +8,7 @@ import test from "ava";
 
 import { z } from "zod";
 
-import { board } from "../../../src/new/recipe-grammar/recipe.js";
+import { board } from "../../../src/new/grammar/board.js";
 
 import { testKit } from "../../helpers/_test-kit.js";
